Add the possibility to exclude URLs which contain certain parts.
It is possible to define a global configuration. This configuration can now be used
to specify an array of url parts, which determine excluded urls. So if a url is enqueued,
the app checks the array of url parts, if at least one of these parts is included in the url.
If it is, the url is excluded from processing and is not added to the queue.